requirements.txt
airflow==1.8.0
alembic==0.8.10
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.48.0
bz2file==0.98
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
croniter==0.3.19
Cython==0.26.1
decorator==4.1.2
dill==0.2.7.1
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
Flask==0.12.2
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.12
funcsigs==1.0.0
functools32==3.2.3.post2
future==0.15.2
futures==3.1.1
gensim==2.3.0
gitdb2==2.0.2
GitPython==2.1.5
google-api-python-client==1.6.4
google-auth==1.1.0
google-cloud-core==0.27.1
google-cloud-storage==1.4.0
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
GoogleAppEngineCloudStorageClient==1.9.22.1
gunicorn==19.3.0
h5py==2.7.1
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
Jinja2==2.8.1
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.8
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
mistune==0.7.4
mock==2.0.0
mysql==0.0.1
MySQL-python==1.2.5
mysqlclient==1.3.12
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
oauth2client==4.1.2
ordereddict==1.1
pandas==0.20.3
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.0.0b2
psutil==4.4.2
psycopg2==2.7.3.1
ptyprocess==0.5.2
pyasn1==0.3.6
pyasn1-modules==0.1.4
Pygments==2.2.0
PyMySQL==0.7.11
python-daemon==2.1.2
python-dateutil==2.6.1
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.18.4
rsa==3.4.2
scandir==1.5
scikit-learn==0.18.2
scipy==0.19.1
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.5.3
smmap2==2.0.3
SQLAlchemy==1.1.14
tabulate==0.7.7
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.6
terminado==0.6
testpath==0.3.1
thrift==0.9.3
tornado==4.5.1
traitlets==4.3.2
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
WTForms==2.1
zope.deprecation==4.3.0
